I made the methods for the 'concertScreen' which are 'listAllConcerts()', 'viewConcertById()', 'addConcert()', 'updateConcert()', 'updateTicketSold()', 'deleteConcert()'. I also made the menu for the 'searchConcertScreen()' and added the methods for it that it will use. In the 'concertService' class I added the methods that the concert screen will use.

+++ src/main/java/com/pluralsight/concerttracker/service/ConcertService.java
@@ -1,8 +1,6 @@
 package com.pluralsight.concerttracker.service;
 
-import com.pluralsight.concerttracker.data.ArtistRepository;
 import com.pluralsight.concerttracker.data.ConcertRepository;
-import com.pluralsight.concerttracker.models.Artist;
 import com.pluralsight.concerttracker.models.Concert;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,17 +17,42 @@ public class ConcertService {
         this.concertRepository = concertRepository;
     }
 
+    //public List<Concert> byYear(int year) { return concertRepository.findByYear(year); }
+
+    public long count(){
+        return concertRepository.count();
+    }
+
     public List<Concert> getAllConcert() {
         return concertRepository.findAll();
     }
+
     public Concert getConcertById(Long id) {
         return concertRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Concert not found."));
     }
-    public Concert saveConcert(Concert concert) {
-        return concertRepository.save(concert);
+    public void saveConcert(Concert concert) {
+        concertRepository.save(concert);
     }
-    public void deleteConcert(Long id) {
+
+    public void updatePrice(long id, double price) {
+        Concert concert = getConcertById(id);
+        concert.setTicket_price(price);
+        concertRepository.save(concert);
+    }
+
+    public void deleteConcert(long id) {
+        if (!concertRepository.existsById(id)) {
+            throw new NotFoundException("No sneaker with id " + id);
+        }
         concertRepository.deleteById(id);
     }
+
+    public void updateTicketSold(long id, int amount) {
+        Concert concert = getConcertById(id);
+        concert.setTicket_sold(amount);
+        concertRepository.save(concert);
+    }
+
+
 }
